refactor : board aggregate segregate comment aggregate and emoji aggregate

+++ src/backend/community-service/src/main/java/harmony/communityservice/board/comment/dto/DeleteCommentRequest.java
@@ -2,6 +2,6 @@ package harmony.communityservice.board.comment.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record DeleteCommentRequest(@NotNull Integer commentId, @NotNull Long userId, @NotNull Long boardId) {
+public record DeleteCommentRequest(@NotNull Long commentId, @NotNull Long userId, @NotNull Long boardId) {
 
 }
